+++ lib/appfuel/storage/aws_dynamodb/repository.rb
@@ -6,6 +6,14 @@ module Appfuel
           "#{super}.aws.dynamodb"
         end
       end
+
+      def storage_class(domain_name)
+        mapper.storage_class('aws.dynamodb', domain_name)
+      end
+
+      def to_entity(domain_name, storage)
+        super(domain_name, 'aws.dynamodb', storage)
+      end
     end
   end
 end